CAD and firmware update
Added option for a yaw-offset for the camera cage
Added option to switch between fixed and manual lens type, set via a parameter in the varset contained in the freecad file.
Added "BCOTI" label on the bikini-cover file
Removed deadzones from the X and Y silders for digital alignment.
Fixed some types.